refactor: move execution engine state to internal types
